# Add NE10 to enable more neon optimization for libopus