--- a/builder/scripts.d/50-libopus.sh
+++ b/builder/scripts.d/50-libopus.sh
@@ -31,9 +31,26 @@ ffbuild_dockerbuild() {
         return -1
     fi
 
+    if [[ $TARGET == linux* || $TARGET == mac* ]] && [[ $TARGET == *arm64 ]]; then
+        myconf+=(
+            --with-NE10-libraries="$FFBUILD_PREFIX"/lib
+            --with-NE10-includes="$FFBUILD_PREFIX"/include/libNE10
+        )
+    fi
+
     ./configure "${myconf[@]}"
     make -j$(nproc)
     make install
+
+    if [[ $TARGET == *arm64 ]]; then
+        if [[ $TARGET == mac* ]]; then
+          gsed -i 's/-lopus/-lopus -lNE10/' "$FFBUILD_PREFIX"/lib/pkgconfig/opus.pc
+          gsed -i 's/-I${includedir}\/opus/-I${includedir}\/opus -I${includedir}\/libNE10/' "$FFBUILD_PREFIX"/lib/pkgconfig/opus.pc
+      else
+          sed -i 's/-lopus/-lopus -lNE10/' "$FFBUILD_PREFIX"/lib/pkgconfig/opus.pc
+          sed -i 's/-I${includedir}\/opus/-I${includedir}\/opus -I${includedir}\/libNE10/' "$FFBUILD_PREFIX"/lib/pkgconfig/opus.pc
+      fi
+    fi
 }
 
 ffbuild_configure() {