Unlike on x86, where libopus provides inline assembly, most of the performance optimizations for ARM Neon are implemented in the NE10 library. We need to build it separately for optimal performance on arm64 targets.

The performance improved significantly on M4 Max using test command:

```
./ffmpeg -f lavfi -i "anoisesrc=d=1000" -c:a libopus -b:a 128k -benchmark -f null -
```

#### Before:

```
[out#0/null @ 0x6000004a8300] video:0KiB audio:14175KiB subtitle:0KiB other streams:0KiB global headers:0KiB muxing overhead: unknown
size=N/A time=00:16:40.00 bitrate=N/A speed=61.7x
bench: utime=16.488s stime=0.590s rtime=16.202s
bench: maxrss=12926976KiB
```

#### After:

```
[out#0/null @ 0x6000035e8000] video:0KiB audio:14171KiB subtitle:0KiB other streams:0KiB global headers:0KiB muxing overhead: unknown
size=N/A time=00:16:40.00 bitrate=N/A speed= 343x    
bench: utime=3.277s stime=0.492s rtime=2.916s
bench: maxrss=11796480KiB
```

It is more than 5 times as fast. Would be more useful on CPU constrained devices like RK3588 based boards.
<!--
Ensure your title is short, descriptive, and in the imperative mood (Fix X, Change Y, instead of Fixed X, Changed Y).
For a good inspiration of what to write in commit messages and PRs please review https://chris.beams.io/posts/git-commit/ and our https://docs.jellyfin.org/general/contributing/issues.html page.
-->

**Changes**
<!-- Describe your changes here in 1-5 sentences. -->

**Issues**
<!-- Tag any issues that this PR solves here.
ex. Fixes # -->