@gnattu 2024-12-20
This would need more work as the NE10 library is very outdated and cannot be compiled with current ct-ng toolchain.

I also suspect the low performance of libopus is actually due to compiler/env bugs in libopus, as the current artifacts generated is still with slow performance.