@nyanmisaka 2024-12-20
libopus AVX2 auto-detection is also broken only in linux portable builds.

// windows & macos amd64/x86
```
2024-12-20T05:06:15.3974267Z       External Assembly Optimizations: 
2024-12-20T05:06:15.3974650Z       Intrinsics Optimizations: ...... x86 SSE SSE2 SSE4.1 AVX2
2024-12-20T05:06:15.3975037Z       Run-time CPU detection: ........ x86 SSE4.1 AVX2
```