@gnattu 2024-12-20
Performance of NE10 on RK3588 built with crosstool-ng:

#### with NE10:

```
[out#0/null @ 0x559f9f9ed0] video:0KiB audio:14172KiB subtitle:0KiB other streams:0KiB global headers:0KiB muxing overhead: unknown
size=N/A time=00:16:40.00 bitrate=N/A speed= 151x    
bench: utime=9.945s stime=0.862s rtime=6.639s
bench: maxrss=16964KiB
```


#### without NE10:

```
[out#0/null @ 0x55bd42ded0] video:0KiB audio:14175KiB subtitle:0KiB other streams:0KiB global headers:0KiB muxing overhead: unknown
size=N/A time=00:16:40.00 bitrate=N/A speed= 135x    
bench: utime=11.357s stime=1.059s rtime=7.402s
bench: maxrss=16816KiB
```

It is like 11.8% performance boost.